clean up

+++ monyy/database.py
@@ -8,11 +8,12 @@ from datetime import date
 from monyy import app, db
 
 
-
 class User(UserMixin, db.Model):
     user_id = db.Column(db.Integer, primary_key=True)
     user_name = db.Column(db.String, unique=True, nullable=False)
     pass_hash = db.Column(db.String, nullable=False)
+
+    # relationships
     accounts = db.relationship('Account', backref=db.backref('User', uselist=False))
 
     def get_id(self):
@@ -29,9 +30,12 @@ class Account(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey(User.user_id), nullable=False)
     account_name = db.Column(db.String, nullable=False)
     account_type = db.Column(db.String, nullable=False)
+
+    # relationships
     transactions = db.relationship('Transaction', backref=db.backref('Account', uselist=True))
     account_tags = db.relationship('Account_tag', backref=db.backref('Account', uselist=True))
     debts = db.relationship('Debt', backref=db.backref('Account', uselist=True))
+
     def __repr__(self):
         return '<Account %r>' % self.account_name
     
@@ -41,33 +45,45 @@ class Transaction(db.Model):
     transaction_type = db.Column(db.String, nullable=False)
     transaction_value = db.Column(db.Float, nullable=False)
     transaction_date = db.Column(db.Date, nullable=False, default=date.today())
-    # transaction_time = db.Column(db.Time, nullable=False)
     transaction_note = db.Column(db.String)
-    transactions_ba = db.relationship('Transaction_bank_account', backref=db.backref('Transaction', uselist=True))
-    transactions_re = db.relationship('Transaction_real_estate', backref=db.backref('Transaction', uselist=True))
-    transactions_bond = db.relationship('Transaction_bond', backref=db.backref('Transaction', uselist=True))
-    transactions_stock = db.relationship('Transaction_stock', backref=db.backref('Transaction', uselist=True))
-    transactions_debt = db.relationship('Transaction_debt', backref=db.backref('Transaction', uselist=True))
-    transaction_tags = db.relationship('Transaction_tag', backref=db.backref('Transaction', uselist=True))
+
+    # relationships
+    transactions_ba = db.relationship(
+        'Transaction_bank_account', backref=db.backref('Transaction', uselist=True))
+    transactions_re = db.relationship(
+        'Transaction_real_estate', backref=db.backref('Transaction', uselist=True))
+    transactions_bond = db.relationship(
+        'Transaction_bond', backref=db.backref('Transaction', uselist=True))
+    transactions_stock = db.relationship(
+        'Transaction_stock', backref=db.backref('Transaction', uselist=True))
+    transactions_debt = db.relationship(
+        'Transaction_debt', backref=db.backref('Transaction', uselist=True))
+    transaction_tags = db.relationship(
+        'Transaction_tag', backref=db.backref('Transaction', uselist=True))
+
     def __repr__(self):
-        return '<Transaction %r>' % self.transaction_type
-    
+        return '<Transaction %r>' % self.transaction_type 
 
 class Bank_account(db.Model):
     bank_account_id = db.Column(db.Integer, primary_key=True)
     bank_name = db.Column(db.String, nullable=False)
     account_digits = db.Column(db.String, nullable=False)
-    interest_rate = db.Column(db.Integer, nullable=True)
+    interest_rate = db.Column(db.Float, nullable=True)
     interest_period = db.Column(db.String, nullable = True)
-    transactions_ba = db.relationship('Transaction_bank_account', backref=db.backref('Bank_account', uselist=False))
+
+    # relationships
+    transactions_ba = db.relationship(
+        'Transaction_bank_account', backref=db.backref('Bank_account', uselist=False))
+
     def __repr__(self):
         return '<Bank Account %r %r>' % (self.bank_name, self.account_digits)
 
 class Transaction_bank_account(db.Model):
     transaction_ba_id =  db.Column(db.Integer, primary_key=True)
-    transaction_id = db.Column(db.Integer, db.ForeignKey(Transaction.transaction_id), nullable=False)
-    bank_account_id = db.Column(db.Integer, db.ForeignKey(Bank_account.bank_account_id), nullable=False)
-
+    transaction_id = db.Column(db.Integer, db.ForeignKey(
+        Transaction.transaction_id), nullable=False)
+    bank_account_id = db.Column(db.Integer, db.ForeignKey(
+        Bank_account.bank_account_id), nullable=False)
 
 class Debt(db.Model):
     debt_id = db.Column(db.Integer, primary_key=True)
@@ -77,79 +93,91 @@ class Debt(db.Model):
     interest_period = db.Column(db.String, nullable = True)
     payment_account = db.Column(db.Integer, db.ForeignKey(Account.account_id), nullable=False)
     payment_date = db.Column(db.String, nullable=True)
-    transactions_debt = db.relationship('Transaction_debt', backref=db.backref('Debt', uselist=False))
+
+    # relationships
+    transactions_debt = db.relationship(
+        'Transaction_debt', backref=db.backref('Debt', uselist=False))
+
     def __repr__(self):
         return '<Debt %r>' % self.lender
 
 class Transaction_debt(db.Model):
     transaction_debt_id = db.Column(db.Integer, primary_key=True)
-    transaction_id = db.Column(db.Integer, db.ForeignKey('transaction.transaction_id'), nullable=False)
+    transaction_id = db.Column(db.Integer, db.ForeignKey(
+        'transaction.transaction_id'), nullable=False)
     debt_id = db.Column(db.Integer, db.ForeignKey('debt.debt_id'), nullable=False)
-    
 
 class Real_estate(db.Model):
     real_estate_id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String, nullable=False)
     original_value = db.Column(db.Float, nullable=False)
     estimated_value = db.Column(db.Float, nullable=False)
-    transactions_re = db.relationship('Transaction_real_estate', backref=db.backref('Real_estate', uselist=False))
+
+    # relationships
+    transactions_re = db.relationship(
+        'Transaction_real_estate', backref=db.backref('Real_estate', uselist=False))
+
     def __repr__(self):
         return '<Real Estate %r>' % self.name
 
 class Transaction_real_estate(db.Model):
     transaction_re_id = db.Column(db.Integer, primary_key=True)
-    transaction_id = db.Column(db.Integer, db.ForeignKey(Transaction.transaction_id), nullable=False)
-    real_estate_id = db.Column(db.Integer, db.ForeignKey(Real_estate.real_estate_id), nullable=False)
-
+    transaction_id = db.Column(db.Integer, db.ForeignKey(
+        Transaction.transaction_id), nullable=False)
+    real_estate_id = db.Column(db.Integer, db.ForeignKey(
+        Real_estate.real_estate_id), nullable=False)
 
 class Bond(db.Model):
     bond_id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String, nullable=False)
     value = db.Column(db.Float, nullable=False)
     maturation_date = db.Column(db.Date, nullable=False, default=date.today())
-    transactions_bond = db.relationship('Transaction_bond', backref=db.backref('Bond', uselist=False))
+
+    # relationships
+    transactions_bond = db.relationship(
+        'Transaction_bond', backref=db.backref('Bond', uselist=False))
+
     def __repr__(self):
         return '<Bond %r>' % self.name
 
 class Transaction_bond(db.Model):
     transaction_bond_id = db.Column(db.Integer, primary_key=True)
-    transaction_id = db.Column(db.Integer, db.ForeignKey(Transaction.transaction_id), nullable=False)
+    transaction_id = db.Column(db.Integer, db.ForeignKey(
+        Transaction.transaction_id), nullable=False)
     bond_id = db.Column(db.Integer, db.ForeignKey(Bond.bond_id), nullable=False)
     
-
 class Stock(db.Model):
     stock_id = db.Column(db.Integer, primary_key=True)
     stock_symbol = db.Column(db.String, nullable=False)
     exchange = db.Column(db.String, nullable=False)
     num_stocks = db.Column(db.Integer, nullable=False) #Starting number of stocks
-    transactions_stock = db.relationship('Transaction_stock', backref=db.backref('Stock', uselist=False))
-    # stock_ids = db.relationship('Stock_value', backref=db.backref('Stock', uselist=True))
+
+    # relationships
+    transactions_stock = db.relationship(
+        'Transaction_stock', backref=db.backref('Stock', uselist=False))
+
     def __repr__(self):
         return '<Stock %r>' % self.stock_symbol
 
-# class Stock_value(db.Model):
-#     stock_value_id = db.Column(db.Float, primary_key=True)
-#     stock_id = db.Column(db.Integer, db.ForeignKey(Stock.stock_id), nullable=False)
-#     date = db.Column(db.Date, nullable=False, default=date.today())
-#     value = db.Column(db.Integer, nullable=False)
-
 class Transaction_stock(db.Model):
     transaction_stock_id = db.Column(db.Integer, primary_key=True)
-    transaction_id = db.Column(db.Integer, db.ForeignKey(Transaction.transaction_id), nullable=False)
+    transaction_id = db.Column(db.Integer, db.ForeignKey(
+        Transaction.transaction_id), nullable=False)
     stock_id = db.Column(db.Integer, db.ForeignKey(Stock.stock_id), nullable=False)
     
-
 class Tag(db.Model):
     tag_id = db.Column(db.Integer, primary_key=True)
     tag_name = db.Column(db.String, unique=True, nullable=False)
     account_tags = db.relationship('Account_tag', backref=db.backref('Tag', uselist=True))
     transaction_tags = db.relationship('Transaction_tag', backref=db.backref('Tag', uselist=True))
+    
     def __repr__(self):
         return '<Tag %r>' % self.tag_name
 
 class Transaction_tag(db.Model):
     transaction_tag_id = db.Column(db.Integer, primary_key=True)
-    transaction_id = db.Column(db.Integer, db.ForeignKey(Transaction.transaction_id), nullable=False)
+    transaction_id = db.Column(db.Integer, db.ForeignKey(
+        Transaction.transaction_id), nullable=False)
     tag_id = db.Column(db.Integer, db.ForeignKey(Tag.tag_id), nullable=False)
     
 class Account_tag(db.Model):
@@ -157,7 +185,6 @@ class Account_tag(db.Model):
     account_id = db.Column(db.Integer, db.ForeignKey(Account.account_id), nullable=False)
     tag_id = db.Column(db.Integer, db.ForeignKey(Tag.tag_id), nullable=False)
     
-
 
 def database_test():
     #Test 1: Add and retrieve values from each table
@@ -171,12 +198,13 @@ def database_test():
     # print(user_selected.get_id())
     # print(user_selected.get_username())
 
-
-    account_test = Account(user_id=user_selected.user_id, account_name='tester', account_type='Savings')
+    account_test = Account(user_id=user_selected.user_id,
+                           account_name='tester', account_type='Savings')
     # print(Account.query.filter_by(account_type='Savings').first())
     db.session.add(account_test)
     db.session.commit()
-    account_test = Account(user_id=user_selected.user_id, account_name='tester2', account_type='Savings')
+    account_test = Account(user_id=user_selected.user_id,
+                           account_name='tester2', account_type='Savings')
     # print(Account.query.filter_by(account_type='Savings').first())
     db.session.add(account_test)
     db.session.commit()  
@@ -188,23 +216,30 @@ def database_test():
     # print(User.query.filter_by(user_id=user_selected.user_id).all())
     # print(Account.query.filter_by(user_id=user_selected.user_id).all())
 
-    transaction_test = Transaction(account_id=account_selected.account_id, transaction_type='Cash', transaction_value=50)
+    transaction_test = Transaction(
+        account_id=account_selected.account_id, transaction_type='Cash', transaction_value=50)
     # print(Transaction.query.filter_by(account_id=1).first())
     db.session.add(transaction_test)
     db.session.commit()
-    transaction_test = Transaction(account_id=account_selected.account_id, transaction_type='Debt', transaction_value=50)
+    transaction_test = Transaction(
+        account_id=account_selected.account_id, transaction_type='Debt', transaction_value=50)
     # print(Transaction.query.filter_by(account_id=1).first())
     db.session.add(transaction_test)
     db.session.commit()
-    transaction_test = Transaction(account_id=account_selected.account_id, transaction_type='Bond', transaction_value=50)
+    transaction_test = Transaction(
+        account_id=account_selected.account_id, transaction_type='Bond', transaction_value=50)
     # print(Transaction.query.filter_by(account_id=1).first())
     db.session.add(transaction_test)
     db.session.commit()
-    transaction_test = Transaction(account_id=account_selected.account_id, transaction_type='Stock', transaction_value=50)
+    transaction_test = Transaction(
+        account_id=account_selected.account_id, transaction_type='Stock', transaction_value=50)
     # print(Transaction.query.filter_by(account_id=1).first())
     db.session.add(transaction_test)
     db.session.commit()
-    transaction_test = Transaction(account_id=account_selected.account_id, transaction_type='Real Estate', transaction_value=50)
+    transaction_test = Transaction(
+        account_id=account_selected.account_id, 
+        transaction_type='Real Estate', 
+        transaction_value=50)
     # print(Transaction.query.filter_by(account_id=1).first())
     db.session.add(transaction_test)
     db.session.commit()
@@ -229,12 +264,15 @@ def database_test():
     bank_account_selected = Bank_account.query.filter_by(bank_name='USAA').first()
     # print(bank_account_selected.bank_name)
     
-    bank_account_trans_test = Transaction_bank_account(transaction_id = transaction_selected_cash.transaction_id, bank_account_id= bank_account_selected.bank_account_id)
+    bank_account_trans_test = Transaction_bank_account(
+        transaction_id=transaction_selected_cash.transaction_id, 
+        bank_account_id=bank_account_selected.bank_account_id)
     # print(Transaction_bank_account.query.filter_by(bank_account_id=1).first())
     db.session.add(bank_account_trans_test)
     db.session.commit()
     # print(Transaction_bank_account.query.get(1))
-    bank_account_trans_selected = Transaction_bank_account.query.filter_by(bank_account_id=1).first()
+    bank_account_trans_selected = Transaction_bank_account.query.filter_by(
+        bank_account_id=1).first()
     # print(bank_account_trans_selected.bank_account_id)
 
 
@@ -247,7 +285,9 @@ def database_test():
     bond_selected = Bond.query.filter_by(name='ex Bond').first()
     # print(bond_selected.maturation_date)
     
-    bond_trans_test = Transaction_bond(transaction_id = transaction_selected_bond.transaction_id, bond_id= bond_selected.bond_id)
+    bond_trans_test = Transaction_bond(
+        transaction_id=transaction_selected_bond.transaction_id, 
+        bond_id=bond_selected.bond_id)
     # print(Transaction_bond.query.filter_by(bond_id=1).first())
     db.session.add(bond_trans_test)
     db.session.commit()
@@ -273,7 +313,9 @@ def database_test():
     # stock_value_selected = Stock_value.query.filter_by(stock_id=1).first()
     # # print(stock_value_selected.date)
     
-    stock_trans_test = Transaction_stock(transaction_id = transaction_selected_stock.transaction_id, stock_id= stock_selected.stock_id)
+    stock_trans_test = Transaction_stock(
+        transaction_id=transaction_selected_stock.transaction_id, 
+        stock_id=stock_selected.stock_id)
     # print(Transaction_stock.query.filter_by(stock_id=1).first())
     db.session.add(stock_trans_test)
     db.session.commit()
@@ -292,7 +334,9 @@ def database_test():
     real_estate_selected = Real_estate.query.filter_by(name='House').first()
     # print(real_estate_selected.name)
     
-    re_trans_test = Transaction_real_estate(transaction_id = transaction_selected_re.transaction_id, real_estate_id= real_estate_selected.real_estate_id)
+    re_trans_test = Transaction_real_estate(
+        transaction_id=transaction_selected_re.transaction_id, 
+        real_estate_id=real_estate_selected.real_estate_id)
     # print(Transaction_real_estate.query.filter_by(real_estate_id=1).first())
     db.session.add(re_trans_test)
     db.session.commit()
@@ -302,7 +346,8 @@ def database_test():
 
 
     #debt
-    debt_test = Debt(lender='Jersey Boys',principal=60,interest_rate=2,interest_period='Month',payment_account=account_selected.account_id, payment_date='Weekly')
+    debt_test = Debt(lender='Jersey Boys', principal=60, interest_rate=2, interest_period='Month',
+                     payment_account=account_selected.account_id, payment_date='Weekly')
     # print(Debt.query.filter_by(lender='Jersey Boys').first())
     db.session.add(debt_test)
     db.session.commit()
@@ -310,7 +355,8 @@ def database_test():
     debt_selected = Debt.query.filter_by(lender='Jersey Boys').first()
     # print(debt_selected.principal)
     
-    debt_trans_test = Transaction_debt(transaction_id = transaction_selected_debt.transaction_id, debt_id= debt_selected.debt_id)
+    debt_trans_test = Transaction_debt(
+        transaction_id=transaction_selected_debt.transaction_id, debt_id=debt_selected.debt_id)
     # print(Transaction_debt.query.filter_by(debt_id=1).first())
     db.session.add(debt_trans_test)
     db.session.commit()
@@ -330,7 +376,9 @@ def database_test():
     tag_selected_two = Tag.query.filter_by(tag_name='personal').first()
     #print(tag_selected.tag_name)
 
-    test_trans_tag = Transaction_tag(transaction_id=transaction_selected_cash.transaction_id,tag_id=tag_selected.tag_id)
+    test_trans_tag = Transaction_tag(
+        transaction_id=transaction_selected_cash.transaction_id, 
+        tag_id=tag_selected.tag_id)
     #print(Transaction_tag.query.filter_by(tag_id=1).first())
     db.session.add(test_trans_tag)
     db.session.commit()
@@ -338,7 +386,8 @@ def database_test():
     tag_trans_selected = Transaction_tag.query.filter_by(tag_id=1).first()
     #print(tag_trans_selected.transaction_id)
 
-    account_tag_test = Account_tag(account_id = account_selected.account_id, tag_id = tag_selected_two.tag_id)
+    account_tag_test = Account_tag(
+        account_id=account_selected.account_id, tag_id=tag_selected_two.tag_id)
     #print(Account_tag.query.filter_by(account_id=1).first())
     db.session.add(account_tag_test)
     db.session.commit()
@@ -346,13 +395,15 @@ def database_test():
     tag_account_selected = Account_tag.query.filter_by(account_id=1).first()
     #print(tag_account_selected.account_id)
 
-    join_test = db.session.query(Account, Transaction, Transaction_bank_account).filter_by(account_id=account_selected.account_id).join(Transaction).join(Transaction_bank_account).first()
+    join_test = db.session.query(Account, Transaction, Transaction_bank_account)
+        .filter_by(account_id=account_selected.account_id)
+        .join(Transaction)
+        .join(Transaction_bank_account)
+        .first()
     #print(join_test)
     #print(join_test.Transaction.transaction_id)
-    sum_test = db.session.query(func.sum(Transaction.transaction_value).label('balance')).filter_by(transaction_type='Cash').first()
-    sum_test = db.session.query(func.sum(Transaction.transaction_value).label('balance')).filter(Transaction.transaction_id<=12).first()
+    sum_test = db.session.query(func.sum(Transaction.transaction_value).label(
+        'balance')).filter_by(transaction_type='Cash').first()
+    sum_test = db.session.query(func.sum(Transaction.transaction_value).label(
+        'balance')).filter(Transaction.transaction_id <= 12).first()
     #print(sum_test.balance)
-
-
-
-#database_test()
